+++ Documents/ReleaseNotes.md
@@ -1,5 +1,10 @@
 ## Release Notes
 
 ### [<- Go Back](../README.md)
-- `11.0.0-Preview-1.1`: `Sandbox` demo updates.
+- `11.0.0-Preview-1.2`: 
+  - `Popup`. Added Show property.
+  - `IAzureMapsPopups`. Added Show method.
+  - `Sandbox`. 
+	- `GeoJSON Demo`. Demonstrates adding layers using GeoJSON files.
+- `11.0.0-Preview-1.1`: `Sandbox`. Demo updates.
 - `11.0.0-Preview-1.0`: Initial pre-release with basic Azure Maps SDK interop scenarios.